+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## master
 
+- Fix crash in presence validator ([@DmitryTsepelev][])
+
 ## 0.1.0 (2019-02-20)
 
-- Initial version. ([@DmitryTsepelev][])
+- Initial version ([@DmitryTsepelev][])
 
 [@DmitryTsepelev]: https://github.com/DmitryTsepelev